backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0

# AI providers
anthropic==0.8.1
google-generativeai==0.3.2
openai==1.3.7
groq==0.4.1

# ML
torch==2.1.0
torch-geometric==2.4.0
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
ctgan==0.7.4
onnx==1.15.0
onnxruntime==1.16.3

# Monitoring
mlflow==2.9.2
evidently

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
